Adding (still in progress) functionality to read/write reconstrctions from/to file. Also preparing for when I do make functionality for other solve methods by how I format my code

+++ SolveScan/CubeTracker/RouxTracker.c
@@ -298,7 +298,7 @@ int resize_moves_applied(RouxTracker* tracker)
         return 1;
     }
 
-    const int new_size = tracker -> MOVES_APPLIED_MAX * 2;
+    const size_t new_size = tracker -> MOVES_APPLIED_MAX * 2;
     char* p = malloc(sizeof(char)* new_size);
     if (!p)
     {
